requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
scikit-learn==1.6.1
python-dotenv==1.1.0
google-genai==1.14.0
streamlit==1.44.1
requests==2.32.3
pydantic==2.11.2